Refactor: Move to endofyear package
